Bug 787502 - post: Merge superclass SyncResponse and subclass SyncStorageResponse into one class SyncStorageResponse.
