+++ src/main/java/org/mozilla/gecko/sync/jpake/stage/GetRequestStage.java
@@ -15,7 +15,7 @@ import org.mozilla.gecko.sync.jpake.JPakeClient;
 import org.mozilla.gecko.sync.net.BaseResource;
 import org.mozilla.gecko.sync.net.BaseResourceDelegate;
 import org.mozilla.gecko.sync.net.Resource;
-import org.mozilla.gecko.sync.net.SyncResponse;
+import org.mozilla.gecko.sync.net.SyncStorageResponse;
 import org.mozilla.gecko.sync.setup.Constants;
 
 import ch.boye.httpclientandroidlib.Header;
@@ -49,7 +49,7 @@ public class GetRequestStage extends JPakeStage {
           Logger.debug(LOG_TAG, "Finished; returning.");
           return;
         }
-        final SyncResponse res = new SyncResponse(response);
+        final SyncStorageResponse res = new SyncStorageResponse(response);
 
         Header etagHeader = response.getFirstHeader("etag");
         if (etagHeader == null) {
